services/worker-video/requirements.txt
openai
structlog
neo4j
moviepy
pillow
pydantic
pydantic-settings
celery[redis]
redis
sqlalchemy[asyncio]
asyncpg
python-dotenv
